shared typings
